Refactor OAuth2 test helpers into setup_oauth2.go with options pattern
Addresses review feedback on PR #228. Extracts OAuth2-specific helpers
out of setup.go into a dedicated setup_oauth2.go, and collapses the
paired `…AsActor` variants into a single options-based API:

- `InitiateOAuth2Connection(t, connectorID, returnTo, opts...)` replaces
  the previous default + `…AsActor` pair.
- `DeliverOAuth2Callback(t, callbackURL, opts...)` replaces the previous
  default + `…AsActor` pair.
- New `WithActor(externalID, namespace)` option carries the actor
  identity into both helpers.

Updates all callers across the OAuth2 integration tests and companion
markdown specs.
